Split create lost report form into 2 components to handle switch
